Fixed some conversion issues of size_type to int by explict (int) casting vector::size

+++ src/Autotune.cpp
@@ -130,7 +130,7 @@ int Autotune::pickParameterClassToUpdate(void)
     int parm = 0;
     for (std::vector<double>::size_type i = 0; i < parWts.size(); i++) {
         if (rn <= parWts[i]) {
-            parm = i;
+            parm = (int)i;
             break;
         }
     }
